feat: Add student authentication and student portal
This commit introduces support for student authentication by adding a password field to the `Student` and `SystemUser` types. It also includes a new `StudentPortal` component and modifies the `App` component to handle authentication flow. The `SystemUser` type now supports the `STUDENT` role. The `updateSystemUser` function was also added to `storageService` to facilitate updating user information.